PP-647 Connections recover after a DB loss
* This options makes the connection pool check for connections before
returning it to the caller to perform jdbc operations. If the
connection fails, then it drops it from the pool and creates a new
one. While this will have slight overhead in running a validation
query before returning the connection, it makes the application
recoverable in the event of database going AWOL. By having this
option, the application resumes operations as normal when the database
is back.