PP-647 Waiting for the Databse to be available when starting up
* Introducing a dropwizard command object that will check for the
* availability of dependent resources. This is to ensure that we
can connect to the database prior to accepting requests for the app
* When starting up the application the ApplicationStartupDependentResourceChecker
will check if the database is available. If unavailable it would wait till
its up. It rechecks during incremental intervals until a connection
can be acquired. When it's successful the command will exit.
* Having the command executed in the docker-startup.sh allows us to wait
till all dependent resources are ready prior to running db migrations
as well as starting the HTTP server to accept requests.
* Also by introducing this removing the sleep 20 in the Dockerfile as it
doesn't make sense doing it now.