PP-1226 move to maintained Alpine OpenJDK baselayer
Pull in bash which is required but not included by default whilst I'm at
it.

Docker's "java" image has been deprecated in favour of the "openjdk"
image and will not receive any updates past 2016-12-31.

Given that the vanilla '8-jre' and '8-jdk' images are Debian Jessie
based and hence bring with them 150MB+ of packages (and hence security
updates to worry about) it seems sensible to move to the much smaller
Alpine images especially now they're part of the offical base layer
builds.