PP-2032: Add database metrics to pay-publicauth
It uses jdbi, so there was already a healthcheck for the database.
Which is why there was not an explicit db healthcheck.

However I needed to customise it to add metrics. Which means we needed a custom
 db healthcheck.

Other than that, this does the same thing as the other apps, it just registers
some metrics and then pulls in the new data for the metrics whenever the healtcheck is called.

This also adds a stats_healthy stat. If it can't get the stats from the database it is 0. If it
is healthy it is 1.