BAU: Clean up logging
* Always log the request ID, even in exceptions.
* Remove commons-lang3 (we never declared it as a dependency and it doesn't get
  us much anyway)

Exception logging changes from:

"Exception - publicauth request - <url> - exception - <msg>"
to
"[<reqId>] Exception - <msg>"

We already log the URL on request start and end, and these log lines contain
the request ID so we can tie the exception to the URL relatively easily.

We shouldn't need to log the name of the service - the logger implementation
can take care of this.

slf4j supports interpolated log messages. We don't need to concatenate strings
in logger calls.