PP-7104 Refactor to use TokenEntity object
Add `TokenEntity` object to represent a `token` retrieved from the
database. For now, insertion doesn't use this entity.

This is to replace returning Maps from the DAO.

Add a `TokenResponse` object to represent the response to requests to
get a token/tokens to replace building the response using a map.