PP-7104 Use toString method of TokenLink to return value
This is so that we can add a model to represent the response to requests
and use the Jackson `ToStringSerializer` to serialize TokenLinks in the
response object.

The existing `toString` method returned the formatted string
'token_link <tokenLinkValue>'. Update log lines/exteption messages to
include the 'token_link ' substring so that it can be removed from the
`toString` method.