PP-5236: use agreed upon api key prefix
Following discussion with the team and leads we'll be using the prefix
`api_${status}_`. This allows us to include prefixes for other external
identifiers (transactions, services, etc.) if we need to and clearly
demonstrates if this is a live or test key.

This can be updated in the future without breaking any of the current
functionality.