BAU — Tell Dependabot to ignore Dropwizard Sentry 4.x
Tell Dependabot to ignore Dropwizard Sentry 4.x because it’s not
going to be a just-merge-the-Dependabot-PR upgrade.

We essentially forked Dropwizard Sentry because it did not support
Dropwizard 3.x — there is now a Dropwizard Sentry 4.x, which
supports Dropwizard 4.x (and maybe Dropwizard 3.x), but we’d need
to do work to go back to using an unmodified version.

Also add comments explaining why we’re ignoring upgrades for
certain Maven dependencies.